solar tracking system
Today in the world of rampant productivity, energy is the
fundamental source upon which the whole civilization is based upon. As it
is said that energy can neither be created nor be destroyed and, in that
response, it can be signified that it can somehow be stored. The attempt
towards making such goal substantiated, this project has been endeavoured
towards unravelling the path of such objectivity. It is quite natural that
constant utilisation of energies somehow opens the door of scarcity as per
as earthly sources are concerned. Sun, in the stand of which, the tallest
source, spiked over for age’s right from the origin of the whole universe,
16
through which life has been conceived, is the basic and the mother source
of all the energies. Considering the very fundamental from the viewpoint
of storing such energy, the project has been unravelled. Energies other than
from the Sun, are the process from which such are been produced through
the burning of various materials, involving emission of a large amount of
pollution, causing the environment and the atmosphere sick day by day.